merge branches

+++ LibraryCPP/queue.h
@@ -5,7 +5,7 @@
 
 // Stores integer values inside
 // Change it to desired type
-typedef double Data;
+typedef int Data;
 
 struct Queue;
